+++ lib/core.dart
@@ -1,3 +1,4 @@
+export 'package:actions_toolkit_dart/src/core/annotation_properties.dart';
 export 'package:actions_toolkit_dart/src/core/core.dart';
 export 'package:actions_toolkit_dart/src/core/exit_code.dart';
 export 'package:actions_toolkit_dart/src/core/input_options.dart';