+++ lib/src/core/utils.dart
@@ -1,5 +1,7 @@
 import 'dart:convert';
 
+import 'annotation_properties.dart';
+
 /// Sanitizes an input into a string so it can be passed into issueCommand safely
 String toCommandValue(Object? input) {
   if (input == null) {
@@ -10,3 +12,17 @@ String toCommandValue(Object? input) {
 
   return jsonEncode(input);
 }
+
+/// Returns a command properties to send with the actual annotation command
+Map<String, Object> toCommandProperties(AnnotationProperties? properties) =>
+    properties == null
+        ? {}
+        : {
+            if (properties.title != null) 'title': properties.title!,
+            if (properties.file != null) 'file': properties.file!,
+            if (properties.startLine != null) 'line': properties.startLine!,
+            if (properties.endLine != null) 'endLine': properties.endLine!,
+            if (properties.startColumn != null) 'col': properties.startColumn!,
+            if (properties.endColumn != null)
+              'endColumn': properties.endColumn!,
+          };