+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * chore: tune GitHub workflow
+* feat: added support for notice annotations and more annotation fields
 
 ## 0.1.0